Add support for GOG version of Skyrim (1.6.659).
There is probably a better way to do this in a way that does not break
non-GOG builds. My guess is another variant similar to SKSEVR would be
the best approach. Unfortunately I'm not a windows user so doing bigger
changes in VS sln files is beyond my grasp. With this patch I can build
JContainers64.dll under Linux/wine that seems to be working fine with
GOG version of the game. Hopefully somebody else can pick it up to
provide a more acceptable solution.